refactor(ui): TE-840 moved `EditableAnomaly` to dto interfaces file
